requirements.txt
Click==7.0
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
python-memcached==1.59
six==1.12.0
SQLAlchemy==1.3.10
Werkzeug==0.16.0
